Add support for tagging builds
Add support for applying a tag when building via the --tag|-t flag. If
supplied, the build will be labelled with the tag.

As an early implementation, this will store the model's blobs _in a
different path_ than the default, matching the tag. For a tag like

  myregistry.com/my-organization/my-repo:my-tag

and a root config path of $HOME/.jozu/, blobs will be stored in

  $HOME/.jozu/storage/myregistry.com/my-organization/my-repo/<store>

This means that building the same model with different tags will
duplicate storage for now; this should be fixed in the future.

Summary of changes in this commit:

* Storage path is no longer .jozuStore appended to whatever config path
  is; instead Storage uses whatever path is provided and build command
  chooses 'storage' subdirecto of the config path
  * This is required to allow all tagged repos to be subdirectories of
    one base 'storage' directory
* If no tag is provided, blobs are stored in $JOZU_HOME/storage/ as
  before (change: storage instead of .jozuStore)
* If a tag is provided, it is applied as an annotation according to the
  OCI spec, using annotation 'org.opencontainers.image.ref.name'
* If a tag is provided but does not have a hostname, the default value
  of localhost is used.