Add commands for pushing and pulling models from remote OCI registries
Add commands

  jmm push <reference>
  jmm pull <reference>

to enable pushing to and pulling from remote registries. Currently, the
implementation is extremely basic, showing no progress indicator or any
output between the beginning and end of a push.

Push and pull have one option, '--http', to enable using an http-only
registry. This is mostly useful for testing locally.