Update existing commands to use credentials, if present
* Update existing commands to use new network options (including
  --tls-verify=false and --plain-http)
* Update all remote registry connections to use credentials store
* Use filepath package instead of path package

This commit drops flags structs from commands, as it is confusing to
have half the configuration defined in one struct (flags) and the other
half in another (options). Instead, flags are bound to options fields
directly. This also allows using the common NetworkOptions struct as an
embedded struct within each command's options.