Fixup issue during build if Kitfile does not define models
Since switching from a list of models to a single model per Kitfile,
Kitfiles that do not define models will deserialize to a Kitfile with an
empty model.

To avoid this, we change the `model` field on Kitfile to a pointer so
that it deserializes to nil when not present.