Rework build process to support both directories and single files
Rework kit build to handle both layers specified as directories and as
regular files. In particular, we need to be able to package and extract
both

  model:
    path: ./model (directory)

and

  model:
    path: ./model.onnx (file)

This requires significant reworking of the build process as previously
we assumed we could always 1) create a directory and 2) extract files
into it.

In addition, this commit changes the build process to stream the build
to a temporary file rather than an in-memory buffer to avoid requiring a
lot of memory.