Add Dockerfile for installing kit as a KServe ClusterStorageContainer
Add a Dockerfile that builds an image suitable to using `kit` as a
ClusterStorageContainer for KServe installs. Included is a short
README.md that gives a general overview of how to use this container.

To come: automating multi-platform builds of this image to allow it to
be more easily used.