Discard zip used for MacOS notarizing; package as .tar.gz instead
Build the zip file that is used for notarizing in a temporary directory
and discard it during the build, as it does not include the license and
readme. Instead, bundle the signed (and sent for notarization) binary in
the normal .tar.gz to include license, readme, etc.

As zip files (and bare executables) cannot be stapled, it should make no
difference how they are bundled for download.