Wrap push targets to display progress bars
Wrap oras.Target in a struct that shows progress bars for the Push
operation. The wrap function can be used on any target to show a
progress bar when we push data to it (note, the oras Copy operation uses
the Pusher interface internally)

This adds a new dependency to go.mod, for github.com/vbauerster/mpb/v8.
The mpb package automatically suppresses logging progress bars when
output is not a TTY.