Fix how directories are handled while packing
Fix some issues with kit pack: there are three "types" of packing we
need to be able to do:

* Single files (e.g. path is `somedir/myFile.txt`)
* Entire directories (e.g. path is `somedir/subdir`)
* The context directory itself (e.g. path is `.`)

For the first case, we create a tarball with only one file in it; for
the second, we compress the directory itself. This allows for the unpack
process to be

1. Create base directory for layer's path (e.g. `somedir` above)
2. Unpack the layer's tarball into the directory

However, treating the third case in the same way would lead to us adding
a directory named `<context-dir>` to the tarball, which is inappropriate
(e.g. if I have my files in /projects/mydir, you want to be able to
unpack to /projects/otherdir).

To avoid this issue (and having to special-case when the path is `.`),
we need to switch the pack and unpack processes to working on relative
paths rather than absolute paths. This updates kit pack and unpack to
work as follows

1. Resolve absolute paths for e.g. Kitfile and context directory
2. Change working directory to the context directory (or target
   directory for unpack)
3. Pack or unpack based on relative paths within the kitfile rather than
   calculated absolute paths
4. If the directory to be packed is `.`, skip creating a directory entry
   for it.

This change requires relaxing some checks for existing directories (i.e.
don't require the --overwrite flag when a directory exists); we still
require the --overwrite flag before overwriting files.