Improve path handling in kit pack to allow intersecting layers
Update how we handle paths to allow layers to intersect, e.g.

model:
  path: ./my-model/
code:
  path: ./my-model/config/

With this change, in addition to processing the ignore file, we also
consider all other layers when packing a layer. If a path is not
excluded by the ignore file, it is still excluded if

1. Some other layer path is a prefix for the current path, and
2. The other layer path is _not_ a prefix for the current layer path

(more explicitly, while processing `./my-model`, we ignore any paths
starting with `./my-model/config`, but when processing
`./my-model/config`, we don't ignore paths starting with `./my-model` as
this would exclude all paths).

This commit refactors how models are handled internally, removing the
Model struct and working on the Kitfile directly instead. Functionality
for ignore files is moved into the filesystem package and extended to
support the test above