Fix flag collision in unpack
The --config flag for unpack (to unpack the Kitfile) collides with the
global --config flag which is used for setting the Kit config directory.

For unpack, the --config flag is parsed as a boolean flag, meaning a
command like

  kit unpack --config ~/.kitops mymodel:mytag

returns a cryptic error about receiving two arguments instead of one
(since ~/.kitops is parsed as a regular argument).

To avoid this, renamed the `--config` flag on kit unpack to `--kitfile`.