Fix path handling when packing layers
Fix an issue where a layer (e.g. model) with a path that includes a
trailing slash results in packing files to the wrong folder. This is
because Go's filepath.Dir will compute the directory differently if
there is a trailing slash vs not:

* dir/my-model/ --> dir is dir/my-model
* dir/my-model  --> dir is dir

To avoid this, use filepath.Clean on paths to remove trailing slashes.

Without this change, a path specified with trailing slash will result in
its files unpacking one directory up the tree.