Add functional tests
Add functional tests that work by calling kit pack/unpack/etc. directly
with generated dummy files in a temporary directory.

Test cases are stored as yaml files matching a test case struct. For
each test case, the test

1. Sets up a temporary directory and subdirectories
2. Configures kit to use this temporary directory for its internal
   storage
3. Generates files from test case within a subdirectory
4. Writes a Kitfile and .kitignore into the subdirectory
5. Packs the directory, tagging the modelkit test:test
6. Unpacks the test:test tag into another directory
7. Verifies that all files from the test case are present, and all
   ignored files are not present
8. Cleans up the temporary directory

Logs for each command executed by the test are saved and printed in case
a failed test, or if the -test.v flag is used.