Add --compression option to kit pack for testing different mediatypes
Add flag --compression to kit pack, with allowed values 'none', 'gzip'
and 'zstd', with a default value of 'gzip'. If 'none' is specified,
modelkit layers are stored as tar files without any compression. If
'zstd' is specified, tarballs are compressed with Zstandard.

Media type are expanded to support plain tar and zstd compression