Implement our own PullModel function to control how storage is used
Add basic implementation for PullModel, which pulls a modelkit from a
remote reference. This is basically equivalent to the Copy() utility
function in the oras library, which uses Fetch() and Push(); however,
this base allows us control over how files are saved and enables
implementing better cleanup and resumable downloads.