Clean up ingest directory on successful pull
If a pull is cancelled via signal (instead of an error), the ingest file
will not be cleaned up. To avoid accrual of partially-downloaded ingest
files (generated by the non-resumable download case), clear the ingest
directory on every successful pull.

This may remove a partially-completed resumable download, but this
should only occur if the user downloads some _other_ modelkit before
attempting to resume the download.