Don't fail remote list operations if repo contains non-modelkits
Normally, attempting to read a manifest from a repo fails if that
descriptor does not refer to a modelkit. However, for the `kit list`
operation, we want to just ignore such errors to allow listing all
modelkits in repositories that may also contain e.g. docker images or
other OCI artifacts.