Fix garbage collection for pulled modelkits
When we pull a modelkit from a remote, we need to ensure its manifest
descriptor is added to both the local index and the shared (non-scoped)
index. Otherwise, garbage collection will fail -- `kit remove
<modelkit>` will not remove the pulled image, since garbage collection
will not kick in on the main backing OCI store.

Calling delete on a non-existent manifest in the store will not trigger
an error, so this issue is hard to notice.