Add --filter flag to unpack command for advanced filtering
Add flag --filter to allow limiting what is unpacked:

* --filter=model,datasets will unpack the model and all datasets
* --filter=datasets:training will unpack only the 'training' dataset

The general format for the filter is

  --filter=<types>:<fields>

where <types> is a comma-separated list of field types (e.g. config,
models, etc.) and <fields> (optional) is an additional comma-separated
list of which layers to unpack. <fields> matches against name and path
(since names are optional) in each Kitfile section.

The --filter flag can also be passed multiple times.