Fix indefinite hang when unpacking over existing files
We have to wait for progress bars to finish before returning in unpack.
However, if the unpack process encounters an error, the progress bar
will never finish, meaning a deferred Wait() call will hang
indefinitely.

Instead, we only wait in cases of success, and return the error if one
is encountered.