Add digest and diff ID to ModelKit config
Add fields 'digest' and 'diffId' to Kitfile fields that represent
ModelKit layers, which are only serialized in JSON format. When the
Kitfile is used as the ModelKit OCI artifact config, it should contain
digests and diff IDs, but when it is deserialized to YAML it should be
the Kitfile as normal.