Refactor layer packing to restructure tarballs
Change how files are packed into tarballs for ModelKit layers: instead
of packing files relative to the layer's path within a Kitfile, we
instead pack them relative to the context directory for the kit command
(i.e. kit pack <path>).

This means that a ModelKit's layers can be extracted directly to
reproduce the files/directories from the original repository, without
relying on the CLI to place files at the correct path.

Additionally, this commit uses UID 1000 for the file owner inside the
tarball instead of 0 (root). This makes it easier to consume ModelKit
layers as layers in an OCI image.