Make unpack process handle both old and new ModelKits
We need to handle existing ModelKits in addition to the new tar format
-- for new ModelKits, we can unpack the tars directly (and simply), but
for older ones, we need to pre-create the layer's directory and
concatenate paths from the tarball.

To avoid overcomplicating things (since we're still relatively early),
we detect new ModelKits via the digest/diffId fields in the config; this
is a lot simpler to support than using new versions on our media types.