Add library for cloning Git repositories locally
Add a library package for cloning Git repositories. Rather than using a
Go-based git library (such as go-git) we just defer to calling the git
CLI directly. This is because libraries are (in my experience)
significantly slower and consume significantly more memory.

For git-lfs clones, we first clone the repository itself, with LFS
disabled, then download the LFS files.