Kitfile generation: read contents of directories to determine type
Read the contents of any directories and attempt to determine their type
for the purposes of Kitfile generation; if a directory contains multiple
files but e.g. they are all 'dataset'-type files, add the directory as a
dataset layer.

This is not a very pretty implementation, but hopefully covers a decent
number of common cases.