Fix race condition when pulling manifest with duplicate digests
When a manifests contains two layers with the same digest, pulling that
ModelKit can fail if we try to download both (identical) layers
simultaneously. The first download to win the race moves the file from
the ingest/ directory to blobs/sha256/, and the second fails as the
source file does not exist.

To avoid this, we pull each digest only once.