Clean up error messages on push failures
We shouldn't really log a tip about "check that you have write
permissions to the organization..." when a push fails unless the push
failed for that reason.

If the returned error is a response error from the registry, log
information about the response code (e.g. status code). Otherwise, just
log the error to point at what went wrong.