Set placeholder values for ModelKit name and author in kit import
To avoid empty values appearing in 'kit list', set name and author from
the passed in repository -- name is the last segment of the URL, author
is the second last. This turns e.g. https://github.com/jozu-ai/kitops
into the following package

  package:
    name: kitops
    authors: [jozu-ai]

An alternative would be to prompt the user (and provide optional flags)
but this a) complicates the command, and b) duplicates the 'edit
Kitfile' functionality in the later step.