Fix bug where we warn of empty layers when progress bars are disabled
When packing a ModelKit, we precompute the total size of files in each
layer so that we can show a progress bar for the pack process. However,
we assume we don't need this size when progress bars are disabled and
simply return a size of 0.

Since the computed size is also used in printing a warning about empty
layers, this leads to us printing a warning that layers are empty when
progress bars are disabled.